backend/requirements.txt
psycopg2-binary==2.9.11
requests==2.31.0
deepdiff==6.3.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6